+++ bin/modules/win/jrivermodule.py
@@ -42,6 +42,7 @@ except ImportError:
 from bin.modules.win.winutils import applicationrunning
 
 _ZONE_RESOLUTION_CACHE = {}
+PLAYLIST_LOOKAHEAD_LIMIT = 64
 
 ###############################################################
 #
@@ -271,23 +272,76 @@ def populate_song_from_fields(song, fields):
     return song
 
 
+def _annotate_current_song_since_last_cortina(playlist, previous_songs):
+    if not playlist:
+        return playlist
+
+    rules = beamSettings.getRules()
+    current_song = playlist[0]
+    current_song.applySongRules(rules)
+    if current_song.IsCortina == 'yes':
+        current_song.SinceLastCortinaCountOverride = 0
+        return playlist
+
+    since_last_cortina_count = 1
+
+    for previous_song in reversed(previous_songs):
+        previous_song.applySongRules(rules)
+        if previous_song.IsCortina == 'yes':
+            break
+        since_last_cortina_count = since_last_cortina_count + 1
+
+    current_song.SinceLastCortinaCountOverride = since_last_cortina_count
+    return playlist
+
+
+def _has_next_tanda_boundary(playlist):
+    if len(playlist) < 2:
+        return False
+
+    rules = beamSettings.getRules()
+    current_song = playlist[0]
+    current_song.applySongRules(rules)
+    seen_cortina = current_song.IsCortina == 'yes'
+
+    for song in playlist[1:]:
+        song.applySongRules(rules)
+        if not seen_cortina:
+            if song.IsCortina == 'yes':
+                seen_cortina = True
+            continue
+
+        if song.IsCortina != 'yes':
+            return True
+
+    return False
+
+
 def read_playlist_from_xml(xml_root, minpos, max_tanda_length):
     playlist = []
-    max_items = get_playlist_fetch_count(max_tanda_length)
+    min_items = get_playlist_fetch_count(max_tanda_length)
+    previous_songs = []
 
     for idx, item in enumerate(xml_root):
-        if idx < minpos:
-            continue
-        if len(playlist) >= max_items:
-            break
-
         fields = {}
         for tag in item:
             fields[tag.attrib.get('Name', '')] = get_xml_text(tag)
 
-        playlist.append(populate_song_from_fields(SongObject(), fields))
+        song = populate_song_from_fields(SongObject(), fields)
 
-    return playlist
+        if idx < minpos:
+            previous_songs.append(song)
+            continue
+
+        playlist.append(song)
+
+        if len(playlist) >= PLAYLIST_LOOKAHEAD_LIMIT:
+            break
+
+        if len(playlist) >= min_items and _has_next_tanda_boundary(playlist):
+            break
+
+    return _annotate_current_song_since_last_cortina(playlist, previous_songs)
 
 
 def call_with_optional_zone(target, method_name, target_zone):
@@ -345,7 +399,7 @@ def run_with_com(target_zone, max_tanda_length):
         raise ImportError()
 
     playlist = []
-    max_items = get_playlist_fetch_count(max_tanda_length)
+    min_items = get_playlist_fetch_count(max_tanda_length)
     pythoncom.CoInitialize()
     try:
         mjautomation = win32com.client.Dispatch("MediaJukebox Application")
@@ -362,8 +416,16 @@ def run_with_com(target_zone, max_tanda_length):
             mjplaylist = call_with_optional_zone(mjautomation, 'GetCurPlaylist', target_zone)
             minpos = int(getattr(mjplaylist, 'Position', 0))
             maxpos = int(mjplaylist.GetNumberFiles())
-            for songpos in range(minpos, min(maxpos, minpos + max_items)):
+            previous_songs = []
+            for songpos in range(0, minpos):
+                previous_songs.append(getSongAt(mjplaylist, songpos))
+            for songpos in range(minpos, maxpos):
                 playlist.append(getSongAt(mjplaylist, songpos))
+                if len(playlist) >= PLAYLIST_LOOKAHEAD_LIMIT:
+                    break
+                if len(playlist) >= min_items and _has_next_tanda_boundary(playlist):
+                    break
+            playlist = _annotate_current_song_since_last_cortina(playlist, previous_songs)
         else:
             playback_status = 'Unknown'
     finally: